+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+* _v0.9.0_  
+       Various function renames. Add -`Rope` suffix to `width`, `insert`,
+       etc and the `append` method of Textual. Makes it more uniform both
+       within the API for Rope and consistent with Map and Set.
+
 * _v0.8.0_  
        Move `write*` functions to Core.Program.Logging. Essentially a
        cosmetic change, but allows us to document the unified output/